// category/src/modules/grpc/proto/category.pb.ts
export interface GetAttributesRequest {
  size: number;
  page: number;
  search: string;
  optionsIncluded?: boolean | undefined;
}

export interface Option {
  id: string;
  nameEn: string;
  nameAr: string;
  status: string;
  positionAr: number;
  positionEn: number;
}
// ...
export interface GetAttributeRequest {
  id: string;
}

export interface GetAttributeResponse {
  attribute: Attribute | undefined;
}

export interface CreateAttributeRequest {
  name: string;
  nameAr: string;
  status: string;
  options: Option[];
}

export interface CreateAttributeResponse {
  id: string;
  name: string;
  nameAr: string;
  status: string;
  options: Option[];
}

export interface GetConditionsRequest {
  ids: string[];
}

export interface GetConditionsResponse {
  conditions: Condition[];
}

export interface GetMultipleAttributeRequest {
  ids: string[];
}